# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for GPTea API.",
        "title": "GPTea API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "api.gptea-test.keenranger.dev",
    "paths": {
...
        "/auth/cred/register": {
            "post": {
                "description": "Register a credential",
                "tags": [
                    "auth"
                ],
                "summary": "Register a credential",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.credBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.messageResponse"
                        }
                    }
                }
            }
        },
        "/auth/cred/sign-in": {
            "post": {
                "description": "Sign in with a credential",
                "tags": [
                    "auth"
                ],
                "summary": "Sign in with a credential",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.credBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.signInHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        },
        "/auth/token/refresh": {
            "post": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    },
                    {
                        "RefreshTokenAuth": []
                    }
                ],
                "description": "Refresh a token",
                "tags": [
                    "token"
                ],
                "summary": "Refresh a token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.signInHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        },
        "/me": {
            "delete": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    }
                ],
                "description": "Delete my account",
                "tags": [
                    "users"
                ],
                "summary": "Delete my account",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        },
        "/me/chats": {
            "get": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    }
                ],
                "description": "Get my chats in descending order of created_at",
                "tags": [
                    "chats"
                ],
                "summary": "Get my chats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.chatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AccessTokenAuth": []
                    }
                ],
                "description": "Post my chat",
                "tags": [
                    "chats"
                ],
                "summary": "Post my chat",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.chatBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.messageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal.Chat": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "Hjejwerhj"
                },
                "name": {
                    "type": "string",
                    "example": "basic"
                }
            }
        },
...
        "server.chatBody": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "server.chatsResponse": {
            "type": "object",
            "properties": {
                "chats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal.Chat"
                    }
                }
            }
        },
        "server.credBody": {
            "type": "object",
            "required": [
                "accessToken",
                "cred"
            ],
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "cred": {
                    "type": "string",
                    "example": "naver"
                }
            }
        },
        "server.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "server.messageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello, World!"
                }
            }
        },
...
        "server.signInHandlerOutput": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AccessTokenAuth": {
            "description": "type `Bearer {access_token}`",
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        },
        "RefreshTokenAuth": {
            "description": "type `{refresh_token}`",
            "type": "apiKey",
            "name": "x-refresh-token",
            "in": "header"
        }
    }
}